src/test/kotlin/me/nicolas/adventofcode/year2024/Day23Test.kt
package me.nicolas.adventofcode.year2024

import org.junit.jupiter.api.Test
import kotlin.test.assertEquals

class Day23Test {
    private val day = Day23(2024, 23)
    
    val test = """
        kh-tc
        qp-kh
        de-cg
        ka-co
        yn-aq
        qp-ub
        cg-tb
        vc-aq
        tb-ka
        wh-tc
        yn-cg
        kh-ub
        ta-co
        de-co
        tc-td
        tb-wq
        wh-td
        ta-ka
        td-qp
        aq-cg
        wq-ub
        ub-vc
        de-ta
        wq-aq
        wq-vc
        wh-yn
        ka-de
        kh-ta
        co-tc
        wh-qp
        tb-vc
        td-yn
    """.trimIndent()

    @Test
    fun partOne() {
        assertEquals(7, day.partOne(test))
    }

    @Test
    fun partTwo() {
        assertEquals("co,de,ka,ta", day.partTwo(test))
    }
}